requirements.txt
modal>=0.61.0
fastapi>=0.115.0
pydantic>=2.9.0
requests>=2.31.0
scenedetect==0.6.1
opencv-python-headless==4.10.0.84
boto3>=1.34.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
